+++ packages/next-example/rewrites.js
@@ -13,17 +13,18 @@ module.exports = [
     // { source: '/connect/:path*', destination: 'http://192.168.64.201:8080/connect/:path*' },
     // { source: '/graphql/:path*', destination: 'http://192.168.64.202:8083/AElfIndexer_DApp/PortKeyIndexerCASchema/graphql/:path*' },
 
-    // // test4
+    // test4
     // { source: '/api/:path*', destination: 'http://192.168.67.179:5001/api/:path*' },
     // { source: '/connect/:path*', destination: 'http://192.168.67.179:8001/connect/:path*' },
     // { source: '/graphql/:path*', destination: 'http://192.168.67.67:8083/AElfIndexer_DApp/PortKeyIndexerCASchema/graphql/:path*' },
 
     // mainnet
     // { source: '/api/:path*', destination: 'https://aa-portkey.portkey.finance/api/:path*' },
-    // { source: '/connect/:path*', destination: 'https://auth-portkey.portkey.finance/connect/:path*' },
+    // { source: '/connect/:path*', destination: 'https://auth-aa-portkey.portkey.finance/connect/:path*' },
+    // { source: '/graphql/:path*', destination: 'https://dapp-aa-portkey.portkey.finance/Portkey_V2_DID/PortKeyIndexerCASchema/graphql/:path*' },
 
     // testnet
     { source: '/api/:path*', destination: 'https://aa-portkey-test.portkey.finance/api/:path*' },
     { source: '/connect/:path*', destination: 'https://auth-aa-portkey-test.portkey.finance/connect/:path*' },
-
+    { source: '/graphql/:path*', destination: 'https://dapp-aa-portkey-test.portkey.finance/Portkey_V2_DID/PortKeyIndexerCASchema/graphql/:path*' },
 ];