+++ packages/next-example/src/app/sign/page.tsx
@@ -18,6 +18,7 @@ import { ChainId } from '@portkey/types';
 import { sleep } from '@portkey/utils';
 import { Button } from 'antd';
 import { FetchRequest } from '@portkey/request';
+import { signalrSell } from '@portkey/socket';
 
 const PIN = '111111';
 let CHAIN_ID: ChainId = 'AELF';
@@ -31,13 +32,14 @@ ConfigProvider.setGlobalConfig({
     },
   },
   loginConfig: {
-    loginMethodsOrder: ['Email', 'Telegram', 'Google', 'Phone', 'Apple', 'Scan'],
+    loginMethodsOrder: ['Email', 'Telegram', 'Google', 'Apple', 'Scan'],
     recommendIndexes: [0, 1],
   },
   requestDefaults: {
     timeout: 30000,
+    baseURL: 'https://aa-portkey-test.portkey.finance',
   },
-  serviceUrl: 'https://localtest-applesign2.portkey.finance',
+  serviceUrl: 'https://aa-portkey-test.portkey.finance',
   /** By default, reCaptcha's siteKey of portkey is used, if it is a self-built service, please use your own siteKey */
   // reCaptchaConfig: {
   //   siteKey: '',
@@ -124,7 +126,7 @@ export default function Sign() {
     <div>
       <div>-----------</div>
       <SignIn
-        // pin={'23aa'}
+        pin={'111111'}
         ref={ref}
         keyboard={true}
         design={design}
@@ -226,7 +228,7 @@ export default function Sign() {
           const wallet = await did.load(PIN);
           console.log('wallet:', wallet);
           // Mock chainId: 'AELF'
-          const result = await did.logout({ chainId: CHAIN_ID }, { onMethod: 'transactionHash' });
+          const result = await did.logout({ chainId: CHAIN_ID });
           console.log(result, 'logout====');
         }}>
         logout