+++ packages/socket/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.0](https://github.com/Portkey-Wallet/portkey-web/compare/v2.0.5...v2.1.0) (2024-03-08)
+
+### Bug Fixes
+
+- 🐛 socket connect ([9f2777f](https://github.com/Portkey-Wallet/portkey-web/commit/9f2777f6673be295bf2ea48671411b34802e0842))
+
+## [2.0.5](https://github.com/Portkey-Wallet/portkey-web/compare/v2.0.4...v2.0.5) (2024-03-04)
+
+**Note:** Version bump only for package @portkey/socket
+
 ## [2.0.4](https://github.com/Portkey-Wallet/portkey-web/compare/v2.0.4-alpha.1...v2.0.4) (2024-03-01)
 
 **Note:** Version bump only for package @portkey/socket